+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 Added:
 
+* Add 'campaigns' sub-module
 * Add support for Node.js v6.x
 
 Changed: